requirements.txt
pygraphviz
replicate
click
python-dotenv
matplotlib
networkx
plotly
toml
flask
flask-cors
requests
Pillow